+++ src/R/dxR/R/dxR-package.R
@@ -4,7 +4,7 @@
 ##' the new DNAnexus platform.
 ##' 
 ##' \tabular{ll}{ Package: \tab dxR\cr Type: \tab Package\cr Version: \tab
-##' 0.392.0\cr License: \tab Apache License (== 2.0)\cr
+##' 0.394.0\cr License: \tab Apache License (== 2.0)\cr
 ##' }
 ##' 
 ##' @name dxR-package